pyproject.toml
[tool.poetry]
name = "cbdetect"
version = "0.1.0"
description = ""
authors = ["Tage Johansson <tajo9185@student.uu.se>"]
readme = "README.md"

[tool.poetry.scripts]
fens = "cbdetect.gen_diagrams:main"
gen_data = "cbdetect.gen_imgs:main"
train = "cbdetect.train_model:main"
square = "cbdetect.inference_scripts:single_square_script"
board = "cbdetect.inference_scripts:single_board_script"
video = "cbdetect.inference_scripts:video_script"
pdf = "cbdetect.inference_scripts:pdf_script"

[tool.poetry.dependencies]
python = "^3.12"
chess = "^1.11.1"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
progressbar2 = "^4.5.0"
numpy = "^1.26.4"
cairosvg = "^2.7.1"
albumentations = "^1.4.23"
torchmetrics = "^1.6.0"
datasets = "^2.21.0"
timm = "^1.0.12"
pycocotools = "^2.0.8"
torch = "^2.5.1"
accelerate = "^0.30.1"
transformers = "^4.47.1"
pymupdf = "^1.25.1"
scikit-image = "^0.24.0"
pymupdf4llm = "^0.0.17"
markdown = "^3.7"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.391"
ruff = "^0.6.1"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
exclude = ["cbdetect/train_model.py"]

[tool.ruff]
exclude = ["cbdetect/train_model.py"]
line-length = 100

[tool.ruff.format]
docstring-code-format = true
preview = true
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff
    "I",  # isort
]

[tool.ruff.lint.isort]
split-on-trailing-comma = false
